Added possibility for end user to submit feedback
Via a form, with simple captcha to avoid spam. Feedback is stored in the
database and an email notification can be sent to the admin email
address in the future - not implemented right now.
An admin interface is added to view and manage feedback submissions.
If the user has provided an email address, the admin can submit a reply
which will be sent to the user via email.